backend: add BlockNaiveAllocator
